add policy to categories , articles , users for each role

+++ app/Http/Controllers/ArticleController.php
@@ -8,6 +8,8 @@ use App\Models\Category;
 use App\Services\ImageService;
 use App\Services\SlugValidationService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -25,6 +27,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create' , Article::class);
         $categories = Category::all();
         return view('articles.create', compact('categories'));
     }
@@ -38,6 +41,7 @@ class ArticleController extends Controller
         if($request->hasFile('image')){
             $validData['image'] = $imageService->upload($request->image , 'articals');
         }
+        $validData['user_id']=Auth::id();
         Article::create($validData);
         return redirect()->route('articles.index')->with('success', 'Article created successfully');
 
@@ -56,6 +60,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('update', $article);
         $categories = Category::all();
         return view('articles.edit', compact('article', 'categories'));
     }
@@ -65,11 +70,13 @@ class ArticleController extends Controller
      */
     public function update(ArticleRequest $request, Article $article, ImageService $imageService)
     {
+        Gate::authorize('update', $article);
         $data = $request->validated();
         if ($request->hasFile('image')) {
             $imageService->delete($article->image);
             $data['image'] = $imageService->upload($request->image, 'articles');
         }
+        $validData['user_id'] = Auth::id();
         $article->update($data);
         return redirect()->route('articles.show' , compact('article'))->with('success', 'Article updated successfully');
     }
@@ -79,6 +86,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article, ImageService $imageService)
     {
+        Gate::authorize('delete', $article);
         $imageService->delete($article->image);
         $article->delete();
         return redirect()->route('articles.index')->with('success', 'Article deleted successfully');
